pyproject.toml
[tool.poetry]
name = "coffeebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
praw = "^7.7.1"
pillow = "^10.2.0"
transformers = "^4.37.1"
accelerate = "^0.26.1"
evaluate = "^0.4.1"
wandb = "^0.16.2"
bitsandbytes = "^0.42.0"
datasets = "^2.16.1"
torchmultimodal-nightly = "^2024.1.31"
# torch = "^2.1.2"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
python-multipart = "^0.0.7"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
